Experimental changes to support 2-way UDP client aka connected UDP.
This was an experiment to see if we can call connect from both sides to
disallow receiving unsolicited UDP traffic. This approach works for one
socket but not for the second as due to some reason, only one socket
receives the traffic. This problem is also described here:
http://forums.codeguru.com/showthread.php?451989-Binding-2-UDP-sockets-on-same-port-connected-to-different-destinations-1-receives

At this point, for UDP, we would continue the server client model,
since only that approach works, even if it has the overhead of checking
the validity of sender for each packet.